requirements.txt
PyGithub
PyInquirer